добавлен механизм пересылки сообщения из канала пользователю

+++ src/main/kotlin/ci/ashamaz/hashtagsubscriber/service/ContactUserService.kt
@@ -3,7 +3,7 @@ package ci.ashamaz.hashtagsubscriber.service
 import ci.ashamaz.hashtagsubscriber.model.ContactUser
 
 interface ContactUserService {
-    fun getCotactUserById(id: Int):ContactUser
+    fun getCotactUserById(id: Long):ContactUser
     fun getCotactUserByUserName(name: String):ContactUser?
     fun getContactUserByChatId(id: Long): ContactUser?
     fun addContactUser(user: ContactUser)
